+++ src/main/java/eu/siacs/conversations/parser/MessageParser.java
@@ -32,6 +32,7 @@ import eu.siacs.conversations.entities.Conversation;
 import eu.siacs.conversations.entities.Conversational;
 import eu.siacs.conversations.entities.Message;
 import eu.siacs.conversations.entities.MucOptions;
+import eu.siacs.conversations.entities.Reaction;
 import eu.siacs.conversations.entities.ReadByMarker;
 import eu.siacs.conversations.entities.ReceiptRequest;
 import eu.siacs.conversations.entities.RtpSessionStatus;
@@ -54,6 +55,8 @@ import im.conversations.android.xmpp.model.Extension;
 import im.conversations.android.xmpp.model.carbons.Received;
 import im.conversations.android.xmpp.model.carbons.Sent;
 import im.conversations.android.xmpp.model.forward.Forwarded;
+import im.conversations.android.xmpp.model.occupant.OccupantId;
+import im.conversations.android.xmpp.model.reactions.Reactions;
 
 public class MessageParser extends AbstractParser implements Consumer<im.conversations.android.xmpp.model.stanza.Message> {
 
@@ -429,7 +432,8 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
         } else if (query != null) {
             Log.d(Config.LOGTAG, account.getJid().asBareJid() + ": received mam result with invalid from (" + original.getFrom() + ") or queryId (" + queryId + ")");
             return;
-        } else if (original.fromServer(account)) {
+        } else if (original.fromServer(account)
+                && original.getType() != im.conversations.android.xmpp.model.stanza.Message.Type.GROUPCHAT) {
             Pair<im.conversations.android.xmpp.model.stanza.Message, Long> f;
             f = getForwardedMessagePacket(original, Received.class);
             f = f == null ? getForwardedMessagePacket(original, Sent.class) : f;
@@ -446,8 +450,10 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
         if (timestamp == null) {
             timestamp = AbstractParser.parseTimestamp(original, AbstractParser.parseTimestamp(packet));
         }
+        final Reactions reactions = packet.getExtension(Reactions.class);
         final LocalizedContent body = packet.getBody();
         final Element mucUserElement = packet.findChild("x", Namespace.MUC_USER);
+        final boolean isTypeGroupChat = packet.getType() == im.conversations.android.xmpp.model.stanza.Message.Type.GROUPCHAT;
         final String pgpEncrypted = packet.findChildContent("x", "jabber:x:encrypted");
         final Element replaceElement = packet.findChild("replace", "urn:xmpp:message-correct:0");
         final Element oob = packet.findChild("x", Namespace.OOB);
@@ -471,11 +477,35 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
             Log.e(Config.LOGTAG, "encountered invalid message from='" + from + "' to='" + to + "'");
             return;
         }
-
-        boolean isTypeGroupChat = packet.getType() == im.conversations.android.xmpp.model.stanza.Message.Type.GROUPCHAT;
         if (query != null && !query.muc() && isTypeGroupChat) {
             Log.e(Config.LOGTAG, account.getJid().asBareJid() + ": received groupchat (" + from + ") message on regular MAM request. skipping");
             return;
+        }
+        final Jid mucTrueCounterPart;
+        final OccupantId occupant;
+        if (isTypeGroupChat) {
+            final Conversation conversation =
+                    mXmppConnectionService.find(account, from.asBareJid());
+            final Jid mucTrueCounterPartByPresence;
+            if (conversation != null) {
+                final var mucOptions = conversation.getMucOptions();
+                occupant = mucOptions.occupantId() ? packet.getExtension(OccupantId.class) : null;
+                final var user =
+                        occupant == null ? null : mucOptions.findUserByOccupantId(occupant.getId());
+                mucTrueCounterPartByPresence = user == null ? null : user.getRealJid();
+            } else {
+                occupant = null;
+                mucTrueCounterPartByPresence = null;
+            }
+            mucTrueCounterPart =
+                    getTrueCounterpart(
+                            (query != null && query.safeToExtractTrueCounterpart())
+                                    ? mucUserElement
+                                    : null,
+                            mucTrueCounterPartByPresence);
+        } else {
+            mucTrueCounterPart = null;
+            occupant = null;
         }
         boolean isMucStatusMessage = InvalidJid.hasValidFrom(packet) && from.isBareJid() && mucUserElement != null && mucUserElement.hasChild("status");
         boolean selfAddressed;
@@ -507,7 +537,7 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
             }
         }
 
-        if ((body != null || pgpEncrypted != null || (axolotlEncrypted != null && axolotlEncrypted.hasChild("payload")) || oobUrl != null) && !isMucStatusMessage) {
+        if (reactions == null && (body != null || pgpEncrypted != null || (axolotlEncrypted != null && axolotlEncrypted.hasChild("payload")) || oobUrl != null) && !isMucStatusMessage) {
             final boolean conversationIsProbablyMuc = isTypeGroupChat || mucUserElement != null || account.getXmppConnection().getMucServersWithholdAccount().contains(counterpart.getDomain().toEscapedString());
             final Conversation conversation = mXmppConnectionService.findOrCreateConversation(account, counterpart.asBareJid(), conversationIsProbablyMuc, false, query, false);
             final boolean conversationMultiMode = conversation.getMode() == Conversation.MODE_MULTI;
@@ -631,8 +661,12 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
             }
             message.markable = packet.hasChild("markable", "urn:xmpp:chat-markers:0");
             if (conversationMultiMode) {
-                message.setMucUser(conversation.getMucOptions().findUserByFullJid(counterpart));
-                final Jid fallback = conversation.getMucOptions().getTrueCounterpart(counterpart);
+                final var mucOptions = conversation.getMucOptions();
+                if (occupant != null) {
+                    message.setOccupantId(occupant.getId());
+                }
+                message.setMucUser(mucOptions.findUserByFullJid(counterpart));
+                final Jid fallback = mucOptions.getTrueCounterpart(counterpart);
                 Jid trueCounterpart;
                 if (message.getEncryption() == Message.ENCRYPTION_AXOLOTL) {
                     trueCounterpart = message.getTrueCounterpart();
@@ -669,9 +703,14 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
                     final boolean trueCountersMatch = replacedMessage.getTrueCounterpart() != null
                             && message.getTrueCounterpart() != null
                             && replacedMessage.getTrueCounterpart().asBareJid().equals(message.getTrueCounterpart().asBareJid());
+                    final boolean occupantIdMatch =
+                            replacedMessage.getOccupantId() != null
+                                    && replacedMessage
+                                            .getOccupantId()
+                                            .equals(message.getOccupantId());
                     final boolean mucUserMatches = query == null && replacedMessage.sameMucUser(message); //can not be checked when using mam
                     final boolean duplicate = conversation.hasDuplicateMessage(message);
-                    if (fingerprintsMatch && (trueCountersMatch || !conversationMultiMode || mucUserMatches) && !duplicate) {
+                    if (fingerprintsMatch && (trueCountersMatch || occupantIdMatch || !conversationMultiMode || mucUserMatches) && !duplicate) {
                         synchronized (replacedMessage) {
                             final String uuid = replacedMessage.getUuid();
                             replacedMessage.setUuid(UUID.randomUUID().toString());
@@ -902,10 +941,14 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
                                             serverMsgId,
                                             timestamp);
                             final Contact contact = account.getRoster().getContact(from);
-                            if (mXmppConnectionService.confirmMessages()
-                                    && !contact.isSelf()
-                                    && remoteMsgId != null
-                                    && contact.showInContactList()) {
+                            // this is the same condition that is found in JingleRtpConnection for
+                            // the 'ringing' response. Responding with delivery receipts predates
+                            // the 'ringing' spec'd
+                            final boolean sendReceipts =
+                                    (mXmppConnectionService.confirmMessages()
+                                                    && contact.showInContactList())
+                                            || Config.JINGLE_MESSAGE_INIT_STRICT_OFFLINE_CHECK;
+                            if (remoteMsgId != null && !contact.isSelf() && sendReceipts) {
                                 processMessageReceipts(account, packet, remoteMsgId, null);
                             }
                         } else if ((query != null && query.isCatchup()) || !offlineMessagesRetrieved) {
@@ -1022,6 +1065,7 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
         final Element displayed = packet.findChild("displayed", "urn:xmpp:chat-markers:0");
         if (displayed != null) {
             final String id = displayed.getAttribute("id");
+            // TODO we don’t even use 'sender' any more. Remove this!
             final Jid sender = InvalidJid.getNullForInvalid(displayed.getAttributeAsJid("sender"));
             if (packet.fromAccount(account) && !selfAddressed) {
                 final Conversation c =
@@ -1047,7 +1091,9 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
                     message = null;
                 }
                 if (message != null) {
+                    // TODO use occupantId to extract true counterpart from presence
                     final Jid fallback = conversation.getMucOptions().getTrueCounterpart(counterpart);
+                    // TODO try to externalize mucTrueCounterpart
                     final Jid trueJid = getTrueCounterpart((query != null && query.safeToExtractTrueCounterpart()) ? mucUserElement : null, fallback);
                     final boolean trueJidMatchesAccount = account.getJid().asBareJid().equals(trueJid == null ? null : trueJid.asBareJid());
                     if (trueJidMatchesAccount || conversation.getMucOptions().isSelf(counterpart)) {
@@ -1082,6 +1128,83 @@ public class MessageParser extends AbstractParser implements Consumer<im.convers
                 }
                 if (displayedMessage != null && selfAddressed) {
                     dismissNotification(account, counterpart, query, id);
+                }
+            }
+        }
+
+        if (reactions != null) {
+            final String reactingTo = reactions.getId();
+            final Conversation conversation =
+                    mXmppConnectionService.find(account, counterpart.asBareJid());
+
+            if (conversation != null) {
+                if (isTypeGroupChat && conversation.getMode() == Conversational.MODE_MULTI) {
+                    final var mucOptions = conversation.getMucOptions();
+                    final var occupantId = occupant == null ? null : occupant.getId();
+                    if (occupantId != null) {
+                        final boolean isReceived = !mucOptions.isSelf(occupantId);
+                        final Message message;
+                        final var inMemoryMessage =
+                                conversation.findMessageWithServerMsgId(reactingTo);
+                        if (inMemoryMessage != null) {
+                            message = inMemoryMessage;
+                        } else {
+                            message =
+                                    mXmppConnectionService.databaseBackend
+                                            .getMessageWithServerMsgId(conversation, reactingTo);
+                        }
+                        if (message != null) {
+                            final var combinedReactions =
+                                    Reaction.withOccupantId(
+                                            message.getReactions(),
+                                            reactions.getReactions(),
+                                            isReceived,
+                                            counterpart,
+                                            mucTrueCounterPart,
+                                            occupantId);
+                            message.setReactions(combinedReactions);
+                            mXmppConnectionService.updateMessage(message, false);
+                        } else {
+                            Log.d(Config.LOGTAG, "message with id " + reactingTo + " not found");
+                        }
+                    } else {
+                        Log.d(
+                                Config.LOGTAG,
+                                "received reaction in channel w/o occupant ids. ignoring");
+                    }
+                } else if (conversation.getMode() == Conversational.MODE_SINGLE) {
+                    final Message message;
+                    final var inMemoryMessage =
+                            conversation.findMessageWithUuidOrRemoteId(reactingTo);
+                    if (inMemoryMessage != null) {
+                        message = inMemoryMessage;
+                    } else {
+                        message =
+                                mXmppConnectionService.databaseBackend.getMessageWithUuidOrRemoteId(
+                                        conversation, reactingTo);
+                    }
+                    final boolean isReceived;
+                    final Jid reactionFrom;
+                    if (packet.fromAccount(account)) {
+                        isReceived = false;
+                        reactionFrom = account.getJid().asBareJid();
+                    } else {
+                        isReceived = true;
+                        reactionFrom = counterpart;
+                    }
+                    packet.fromAccount(account);
+                    if (message != null) {
+                        final var combinedReactions =
+                                Reaction.withFrom(
+                                        message.getReactions(),
+                                        reactions.getReactions(),
+                                        isReceived,
+                                        reactionFrom);
+                        message.setReactions(combinedReactions);
+                        mXmppConnectionService.updateMessage(message, false);
+                    } else {
+                        Log.d(Config.LOGTAG, "message with id " + reactingTo + " not found");
+                    }
                 }
             }
         }