+++ src/main/java/eu/siacs/conversations/services/XmppConnectionService.java
@@ -115,6 +115,7 @@ import eu.siacs.conversations.entities.MucOptions;
 import eu.siacs.conversations.entities.MucOptions.OnRenameListener;
 import eu.siacs.conversations.entities.Presence;
 import eu.siacs.conversations.entities.PresenceTemplate;
+import eu.siacs.conversations.entities.Reaction;
 import eu.siacs.conversations.entities.Roster;
 import eu.siacs.conversations.entities.ServiceDiscoveryResult;
 import eu.siacs.conversations.generator.AbstractGenerator;
@@ -156,7 +157,6 @@ import eu.siacs.conversations.xml.LocalizedContent;
 import eu.siacs.conversations.xml.Namespace;
 import eu.siacs.conversations.xmpp.InvalidJid;
 import eu.siacs.conversations.xmpp.Jid;
-import eu.siacs.conversations.xmpp.OnBindListener;
 import eu.siacs.conversations.xmpp.OnContactStatusChanged;
 import eu.siacs.conversations.xmpp.OnKeyStatusUpdated;
 import eu.siacs.conversations.xmpp.OnMessageAcknowledged;
@@ -1251,7 +1251,11 @@ public class XmppConnectionService extends Service {
         }
 
         final PowerManager powerManager = getSystemService(PowerManager.class);
-        this.wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "Conversations:Service");
+        if (powerManager != null) {
+            this.wakeLock =
+                    powerManager.newWakeLock(
+                            PowerManager.PARTIAL_WAKE_LOCK, "Conversations:Service");
+        }
 
         toggleForegroundService();
         updateUnreadCountBadge();
@@ -1837,7 +1841,7 @@ public class XmppConnectionService extends Service {
         sendIqPacket(account, retrieve, (response) -> {
             if (response.getType() == Iq.Type.RESULT) {
                 final Element pubsub = response.findChild("pubsub", Namespace.PUBSUB);
-                final Map<Jid, Bookmark> bookmarks = Bookmark.parseFromPubsub(pubsub, account);
+                final Map<Jid, Bookmark> bookmarks = Bookmark.parseFromPubSub(pubsub, account);
                 processBookmarksInitial(account, bookmarks, true);
             }
         });
@@ -4628,6 +4632,53 @@ public class XmppConnectionService extends Service {
                 item,
                 itemId.toEscapedString(),
                 PublishOptions.persistentWhitelistAccessMaxItems());
+    }
+
+    public boolean sendReactions(final Message message, final Collection<String> reactions) {
+        if (message.getConversation() instanceof Conversation conversation) {
+            final String reactToId;
+            final Collection<Reaction> combinedReactions;
+            if (conversation.getMode() == Conversational.MODE_MULTI) {
+                final var self = conversation.getMucOptions().getSelf();
+                final String occupantId = self.getOccupantId();
+                if (Strings.isNullOrEmpty(occupantId)) {
+                    Log.d(Config.LOGTAG, "occupant id not found for reaction in MUC");
+                    return false;
+                }
+                reactToId = message.getServerMsgId();
+                combinedReactions =
+                        Reaction.withOccupantId(
+                                message.getReactions(),
+                                reactions,
+                                false,
+                                self.getFullJid(),
+                                conversation.getAccount().getJid(),
+                                occupantId);
+            } else {
+                if (message.isCarbon() || message.getStatus() == Message.STATUS_RECEIVED) {
+                    reactToId = message.getRemoteMsgId();
+                } else {
+                    reactToId = message.getUuid();
+                }
+                combinedReactions =
+                        Reaction.withFrom(
+                                message.getReactions(),
+                                reactions,
+                                false,
+                                conversation.getAccount().getJid());
+            }
+            if (Strings.isNullOrEmpty(reactToId)) {
+                return false;
+            }
+            final var reactionMessage =
+                    mMessageGenerator.reaction(conversation, reactToId, reactions);
+            sendMessagePacket(conversation.getAccount(), reactionMessage);
+            message.setReactions(combinedReactions);
+            updateMessage(message, false);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public MemorizingTrustManager getMemorizingTrustManager() {